Create script to run all relevant (7) cases
Created bash script that reads the
config.yaml file to create a sbatch
script. This is then used to run all of
the 7 relevant cases. Adjustments
regarding hyperparameter can be
performed by changing the corresponding
values in config.yaml.
NOTE: If "case" is present in the config
passed to main, the "case" field in the
respective config file is used over the cmd
argument "case".